% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflactar_DA.R
\name{deflactar_DA}
\alias{deflactar_DA}
\title{Llevar a precios constantes las bases}
\usage{
deflactar_DA(
  data,
  mes_base,
  variables_monetarias = c("p10", "p25", "w_mean", "w_median", "p75", "p90", "p99"),
  pisar_datos = F
)
}
\arguments{
\item{data}{Dataframe que se quiere deflactar/inflar}

\item{mes_base}{Mes que se quiere utilizar como base para inflar/deflactar. Formato: "YYYY-MM-DD"}

\item{variables_monetarias}{Se indica mediante un vector las variables monetarias que se quieren deflactar en la base. Por default se utilizan todas las variables monetarias, en caso de querer deflactar algunas específicas indicarlo en la función.}
}
\value{
A matrix of the infile
}
\description{
Se puede inflar o deflactar los datos con la función.
Dados los problemas en la medición de la inflación durante 2007-2015 la base es una construcción propia del CEPXXI
Todos los datos fueron provistos por el Centro de Estudios para la Produccion XXI
}
