% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga_DA.R
\name{descarga_DA}
\alias{descarga_DA}
\title{Funciones para descargar Datos de Datos Abiertos}
\arguments{
\item{index_base}{Valor que toma la columna en el dataset que se quiere descargar}

\item{genero}{Puede tomar valor SI o NO}

\item{jurisdiccion}{Puede tomar los siguientes valores: NO, Departamento vivienda, Provincia trabajo, Departamento fiscal, Provincia vivienda}

\item{universo}{Puede tomar los siguientes valores: NO, Privado, Total empresas, Público, Total empleo}

\item{sector}{Puede tomar los siguientes valores: NO, Letra, CLAE2, CLAE3, CLAE6}

\item{tipo}{Puede tomar los siguientes valores: Percentil Salario, Salario Mediano, Salario prom por depto y sector, Puestos provincia y sector, Salario mediano por provincia y sector, Var. mediana facturación, Cant empleadoras, Puestos depto, Salario promedio, Salario mediano por depto y sector, Salario prom por provincia y sector, \% Muj provincia y sector, Puestos por sector, \% Muj por sector}

\item{show_info_DA}{Valor T/F. Muestra el nombre de la base descargada, su descripción y el link de acceso a la metodología}
}
\value{
A matrix of the infile
}
\description{
Esta funcion sirve para cargar los posibles cruces que pueden
realizarse con la informacion disponible en Datos Abiertos.
Todos los datos fueron provistos por el Centro de Estudios para la Produccion XXI
}
