% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga_DP.R
\name{descarga_DP}
\alias{descarga_DP}
\title{Funciones para descargar Datos de Datos Productivos}
\usage{
descarga_DP(index, show_data_info = T)
}
\arguments{
\item{index}{Valor que toma la columna id en el dataset DPexistentes para la serie requerida}

\item{show_data_info}{Valor T/F. Muestra el nombre de la base descargada, su descripción y su fuente}
}
\value{
A matrix of the infile
}
\description{
Esta funcion sirve para cargar los indicadores disponibles en diferentes fuentes de datos relevadas por el CEP
Todos los datos fueron relevados por el Centro de Estudios para la Produccion XXI
}
