% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicc_depto_prov.R, R/diccionario_depto_y_prov.R
\name{dicc_depto_prov}
\alias{dicc_depto_prov}
\title{Añadir etiquetas a departamento y provincia}
\usage{
dicc_depto_prov(data)

dicc_depto_prov(data)
}
\arguments{
\item{data}{Nombre del elemento al que se quiere añadir las etiquetas}
}
\value{
A matrix of the infile

A matrix of the infile
}
\description{
Esta función sirve para añadir las etiquetas a los departamentos y provincias de los datos descargados.
Todos los datos fueron provistos por el Centro de Estudios para la Produccion XXI

Esta función sirve para añadir las etiquetas a los departamentos y provincias de los datos descargados.
Todos los datos fueron provistos por el Centro de Estudios para la Produccion XXI
}
