% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diccionario_sectores.R
\name{diccionario_sectores}
\alias{diccionario_sectores}
\title{Añadir etiquetas a sector}
\usage{
diccionario_sectores(
  data,
  agregacion_deseada = c("clae6", "clae3", "clae2", "letra"),
  descripciones = T
)
}
\arguments{
\item{data}{Nombre del elemento al que se quiere añadir las etiquetas}

\item{agregacion_deseada}{Vector con los sectores que se quieren añadir. Posibles opciones: c("clae6","clae3","clae2","letra"). Default = todas las agregaciones}

\item{descripciones}{Parámetro T/F: se indica si se quieren agregar la descripción de cada sector o no. Default == T}
}
\value{
A matrix of the infile
}
\description{
Esta función sirve para añadir las etiquetas a los sectores de los datos descargados.
A su vez, permite añadir agregaciones mayores. De esta forma se pueden agregar los datos a otro nivel o bien filtrar la información por algún sector en particular.
Todos los datos fueron provistos por el Centro de Estudios para la Produccion XXI
}
