% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPexistentes.R
\name{DPexistentes}
\alias{DPexistentes}
\title{Funciones para conocer las series disponibles en Datos Productivos}
\usage{
DPexistentes()
}
\value{
A matrix of the infile
}
\description{
Se carga un dataset con los datos disponibles en Datos Productivos
Para conocer los datos y visualizar: https://indicadoresargentina.produccion.gob.ar/inicio
Todos los datos fueron relevados por el Centro de Estudios para la Produccion XXI
}
