% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablero_provincial.R
\name{tablero_provincial}
\alias{tablero_provincial}
\title{Carga tablero Shiny provincial}
\usage{
tablero_provincial()
}
\value{
A matrix of the infile
}
\description{
La función se utiliza para correr el script que abre un tablero dinámico en shiny.
}
