% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflactar_DA.R
\name{deflactar_DA}
\alias{deflactar_DA}
\title{Llevar a precios constantes las bases}
\usage{
deflactar_DA(data, mes_base)
}
\arguments{
\item{data}{Dataframe que se quiere deflactar/inflar}

\item{mes_base}{Mes que se quiere utilizar como base para inflar/deflactar. Formato: "YYYY-MM-DD"}
}
\value{
A matrix of the infile
}
\description{
Se puede inflar o deflactar los datos con la funci贸n.
Dados los problemas en la medici贸n de la inflaci贸n durante 2007-2015 la base es una construcci贸n propia del CEPXXI
Todos los datos fueron provistos por el Centro de Estudios para la Produccion XXI
}
