% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dicc_depto_prov.R
\name{dicc_depto_prov}
\alias{dicc_depto_prov}
\title{A침adir etiquetas a departamento y provincia}
\usage{
dicc_depto_prov(data)
}
\arguments{
\item{data}{Nombre del elemento al que se quiere a침adir las etiquetas}
}
\value{
A matrix of the infile
}
\description{
Esta funci칩n sirve para a침adir las etiquetas a los departamentos y provincias de los datos descargados.
Todos los datos fueron provistos por el Centro de Estudios para la Produccion XXI
}
